+++ testing/test_multi.py
@@ -223,8 +223,9 @@ class TestGroup:
         group.terminate(1.0)
 
 
+@pytest.mark.xfail(reason="active_count() has been broken for some time")
 def test_safe_terminate(execmodel):
-    if execmodel.backend != "threading":
+    if execmodel.backend not in ("thread", "main_thread_only"):
         pytest.xfail(
             "execution model %r does not support task count" % execmodel.backend
         )
@@ -246,8 +247,9 @@ def test_safe_terminate(execmodel):
     assert execmodel.active_count() == active
 
 
+@pytest.mark.xfail(reason="active_count() has been broken for some time")
 def test_safe_terminate2(execmodel):
-    if execmodel.backend != "threading":
+    if execmodel.backend not in ("thread", "main_thread_only"):
         pytest.xfail(
             "execution model %r does not support task count" % execmodel.backend
         )